fix: 앱 컴포넌트에서 jsx 구문 제거
